Split font-related utils/constants out into respective modules
