+++ src/_includes/components/Preview/constants.js
@@ -1,0 +1,1 @@
+export const defaultFonts = ['Inter'];
